pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pants-example"
version = "0.1.0"
description = "Pants example repo"
authors = ["Nick Grisafi <njgrisafi@gmail.com>"]
readme = "README.md"
repository = "https://github.com/njgrisafi/pants-example"
packages = [
    {include = "app"}
]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"

[tool.poetry.dev-dependencies]
"pantsbuild.pants" = "^2.14.0"
